+++ nostr_relay/cli.py
@@ -35,8 +35,14 @@ def main(ctx, config):
     default=False,
     help="Use uvicorn instead of gunicorn",
 )
+@click.option(
+    "--use-purple",
+    is_flag=True,
+    default=False,
+    help="Use purple instead of gunicorn",
+)
 @click.pass_context
-def serve(ctx, use_uvicorn):
+def serve(ctx, use_uvicorn, use_purple):
     """
     Start the http relay server
     """
@@ -54,6 +60,10 @@ def serve(ctx, use_uvicorn):
         from .web import run_with_uvicorn
 
         run_with_uvicorn()
+    elif use_purple:
+        from .purple import serve
+
+        serve(Config)
     else:
         from .web import run_with_gunicorn
 
@@ -167,12 +177,12 @@ async def update_tags(ctx, query):
     Update the tags in the tag table, from a REQ query
     """
     from .storage import get_storage
-    from .util import json
+    from .util import json_loads
 
     if not query:
         query = '[{"since": 1}]'
 
-    query = json.loads(query)
+    query = json_loads(query)
 
     async with get_storage() as storage:
         count = 0
@@ -192,12 +202,12 @@ async def reverify(ctx):
     Reverify all NIP-05 metadata events
     """
     from .storage import get_storage
-    from .util import json
+    from .util import json_loads
 
     async with get_storage() as storage:
         count = 0
         async for event in storage.run_single_query([{"kinds": [0]}]):
-            meta = json.loads(event.content)
+            meta = json_loads(event.content)
             if "nip05" in meta:
                 try:
                     await storage.verifier.verify(event)
@@ -249,9 +259,7 @@ async def load(ctx, filename):
     else:
         fileobj = sys.stdin
     import collections
-    from .util import json
-
-    loads = json.loads
+    from .util import json_loads
 
     Config.authentication["enabled"] = False
     # this will reverify profiles but not reject unverified events
@@ -272,7 +280,7 @@ async def load(ctx, filename):
             line = fileobj.readline()
             if not line:
                 break
-            js = loads(line)
+            js = json_loads(line)
             if isinstance(js, list):
                 event = js[1]
             else:
@@ -410,13 +418,13 @@ async def purge(ctx, query, authors):
     """
     Purge events from the specified query
     """
-    from .util import json
+    from .util import json_loads
     from .storage import get_storage
 
     if authors:
         query = {"authors": [a.strip() for a in authors.readlines()]}
     else:
-        query = json.loads(query)
+        query = json_loads(query)
     print(repr(query))
     async with get_storage() as storage:
         events = []
@@ -493,3 +501,46 @@ async def reindex(since, until, batch):
 
 
 main.add_command(fts)
+
+
+@main.command()
+@async_cmd
+async def bench():
+    from nostr_relay.storage import get_storage
+    from nostr_relay.storage import kv
+    from nostr_relay.util import easy_profiler
+
+    import time
+
+    import logging
+
+    logging.basicConfig(level="INFO")
+
+    async with get_storage() as storage:
+        plan = kv.planner([{"kinds": [1], "limit": 50}])[0]
+        print(plan)
+
+        async def do_queries():
+            queries = 0
+            events = 0
+            start = time.perf_counter()
+            stop = start + 5
+
+            while time.perf_counter() < stop:
+                # for event in kv.execute_one_plan(storage.db, plan, storage.log)[1]:
+                #     events += 1
+                # queries += 1
+                async for event in storage.run_single_query(
+                    [{"kinds": [1], "limit": 25}]
+                ):
+                    events += 1
+                queries += 1
+
+            duration = time.perf_counter() - start
+            print(f"qps: {queries/duration:.1f}  events/sec: {events/duration:.1f}")
+
+        # with easy_profiler():
+        tasks = [asyncio.create_task(do_queries()) for i in range(3)]
+
+        # tasks = [asyncio.get_running_loop().run_in_executor(None, do_queries) for i in range(4)]
+        await asyncio.wait(tasks)