+++ nostr_relay/config.py
@@ -42,6 +42,9 @@ class ConfigClass:
                 }
             },
             "loggers": {
+                "websockets.server": {
+                    "level": "WARNING",
+                },
                 "sqlalchemy.engine": {
                     "level": "WARNING",
                 },
@@ -83,6 +86,7 @@ class ConfigClass:
 
             multiprocessing.current_process().name = proc_name
         self._is_loaded = True
+        self.config_file = filename
 
     def __getattr__(self, attrname):
         return None